requirements.txt
streamlit
python-dotenv
langchain
langgraph
langchain-openai
langchain-community
tavily-python
pydantic
reportlab
